refactor: 샘플 액티비티 코드 분리
