+++ app/src/main/java/woowacourse/payments/ui/features/addcard/AddCardScreen.kt
@@ -1,0 +1,39 @@
+package woowacourse.payments.ui.features.addcard
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import woowacourse.payments.ui.theme.AndroidpaymentsTheme
+
+@Composable
+fun AddCardScreen(onNavigateBack: () -> Unit) {
+    Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+        Greeting(
+            name = "Android",
+            modifier = Modifier.padding(innerPadding),
+        )
+    }
+}
+
+@Composable
+fun Greeting(
+    name: String,
+    modifier: Modifier = Modifier,
+) {
+    Text(
+        text = "Hello $name!",
+        modifier = modifier,
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun GreetingPreview() {
+    AndroidpaymentsTheme {
+        Greeting("Android")
+    }
+}
